requirements.txt
django
django-cors-headers
Pillow
openai
transformers
ipdb
opencv-python
